+++ spec/unit/installer/analyzer_spec.rb
@@ -1104,9 +1104,11 @@ module Pod
             source SpecHelper.test_repo_url
             platform :ios, '6.0'
             project 'Sample Extensions Project/Sample Extensions Project'
+            pod 'matryoshka/Bar'
 
             target 'Sample Extensions Project' do
               pod 'JSONKit', '1.4'
+              pod 'matryoshka/Foo'
             end
 
             target 'Today Extension' do
@@ -1122,17 +1124,23 @@ module Pod
 
           result.targets.flat_map { |at| at.pod_targets.map { |pt| "#{at.name}/#{pt.name}" } }.sort.should == [
             'Pods-Sample Extensions Project/JSONKit',
+            'Pods-Sample Extensions Project/matryoshka-Bar-Foo',
             'Pods-Sample Extensions Project/monkey',
+            'Pods-Today Extension/matryoshka-Bar',
             'Pods-Today Extension/monkey',
           ].sort
           result.targets.flat_map { |at| at.pod_targets_for_build_configuration('Debug').map { |pt| "#{at.name}/Debug/#{pt.name}" } }.sort.should == [
             'Pods-Sample Extensions Project/Debug/JSONKit',
+            'Pods-Sample Extensions Project/Debug/matryoshka-Bar-Foo',
             'Pods-Sample Extensions Project/Debug/monkey',
+            'Pods-Today Extension/Debug/matryoshka-Bar',
             'Pods-Today Extension/Debug/monkey',
           ].sort
           result.targets.flat_map { |at| at.pod_targets_for_build_configuration('Release').map { |pt| "#{at.name}/Release/#{pt.name}" } }.sort.should == [
             'Pods-Sample Extensions Project/Release/JSONKit',
             'Pods-Sample Extensions Project/Release/monkey',
+            'Pods-Sample Extensions Project/Release/matryoshka-Bar-Foo',
+            'Pods-Today Extension/Release/matryoshka-Bar',
             'Pods-Today Extension/Release/monkey',
           ].sort
         end
@@ -1143,6 +1151,8 @@ module Pod
 
           result.targets.flat_map { |at| at.pod_targets.map { |pt| "#{at.name}/#{pt.name}" } }.sort.should == [
             'Pods-Sample Extensions Project/JSONKit',
+            'Pods-Sample Extensions Project/matryoshka-Bar-Foo',
+            'Pods-Today Extension/matryoshka-Bar',
             'Pods-Today Extension/monkey',
           ].sort
         end